move subproduct_pre_save to signals.py
